# openapi.yaml
openapi: 3.0.0
info:
  title: ZENU - AI Agent for Radio Presenters API
  description: |
    ZENU is an AI-powered digital producer for radio presenters and OAPs.
    This API provides digital skills for IBM watsonx Orchestrate integration.
    
    **Core Workflow:**
    Fetch News → Clean & Summarize → Rewrite in OAP Tone → Build Script → Deliver to Teleprompter
  version: 1.0.0
  contact:
    name: ZENU Support
    email: support@zenu.app
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://zenu.app/api
    description: Production server
...
paths:
...
  /skills/fetch-news:
    post:
      tags:
        - News Skills
      summary: "Digital Skill 1: Fetch News from Multiple Sources"
      description: |
        Fetches news from configured sources (RSS feeds, APIs, web scraping).
        This is the first skill in the ZENU workflow orchestration.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sources:
                  type: array
                  items:
                    type: string
                  example: ["vanguard", "punch", "arise", "local"]
                categories:
                  type: array
                  items:
                    type: string
                  example: ["politics", "entertainment", "sports"]
                limit:
                  type: integer
                  default: 10
                  example: 10
      responses:
        '200':
          description: News fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RawNewsItem'

  # ==================== DIGITAL SKILL 2: CLEAN & SUMMARIZE ====================
  /skills/clean-news:
    post:
      tags:
        - News Skills
      summary: "Digital Skill 2: Clean & Summarize News"
      description: |
        Removes ads, noise, and irrelevant content from raw news.
        Generates clean summaries suitable for broadcast.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rawNews
              properties:
                rawNews:
                  type: array
                  items:
                    $ref: '#/components/schemas/RawNewsItem'
      responses:
        '200':
          description: News cleaned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CleanedNewsItem'

  # ==================== DIGITAL SKILL 3: REWRITE IN OAP TONE ====================
  /skills/rewrite-tone:
    post:
      tags:
        - Style Skills
      summary: "Digital Skill 3: Rewrite News in OAP's Tone"
      description: |
        Uses presenter profile to rewrite news in their exact speaking style,
        including language preferences, cadence, and signature phrases.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cleanedNews
                - presenterId
              properties:
                cleanedNews:
                  type: array
                  items:
                    $ref: '#/components/schemas/CleanedNewsItem'
                presenterId:
                  type: string
                  example: "user123"
      responses:
        '200':
          description: News rewritten successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/StyledNewsItem'

  # ==================== DIGITAL SKILL 4: POPULATE SHOW TEMPLATE ====================
  /skills/populate-template:
    post:
      tags:
        - Format Skills
      summary: "Digital Skill 4: Populate Show Template"
      description: |
        Injects styled news into the presenter's pre-defined show structure
        (intro, weather, trending news, global headlines, etc.)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - styledNews
                - presenterId
              properties:
                styledNews:
                  type: array
                  items:
                    $ref: '#/components/schemas/StyledNewsItem'
                presenterId:
                  type: string
                templateOverride:
                  type: object
      responses:
        '200':
          description: Template populated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/PopulatedScript'

  # ==================== DIGITAL SKILL 5: GENERATE FINAL SCRIPT ====================
  /skills/generate-script:
    post:
      tags:
        - Format Skills
      summary: "Digital Skill 5: Generate Final Teleprompter Script"
      description: |
        Formats the populated template into teleprompter-ready chunks
        with timing, emphasis markers, and pause cues.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - populatedScript
              properties:
                populatedScript:
                  $ref: '#/components/schemas/PopulatedScript'
                formatting:
                  type: object
                  properties:
                    fontSize:
                      type: string
                      default: "large"
                    scrollSpeed:
                      type: string
                      default: "medium"
      responses:
        '200':
          description: Script generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/TeleprompterScript'

  # ==================== DIGITAL SKILL 6: DELIVER SCRIPT ====================
...
  /skills/complete-workflow:
    post:
      tags:
        - Style Skills
      summary: Execute Complete ZENU Workflow
      description: |
        Orchestrates all 6 digital skills in sequence:
        Fetch → Clean → Rewrite → Populate → Generate → Deliver
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                sources:
                  type: array
                  items:
                    type: string
                deliveryMethod:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Workflow completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  scriptId:
                    type: string
                  teleprompterUrl:
                    type: string

# ==================== COMPONENTS ====================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RawNewsItem:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        title:
          type: string
        content:
          type: string
        url:
          type: string
        publishedAt:
          type: string
          format: date-time
    CleanedNewsItem:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        title:
          type: string
        summary:
          type: string
        category:
          type: string
        relevanceScore:
          type: number
    StyledNewsItem:
      type: object
      properties:
        id:
          type: string
        originalTitle:
          type: string
        styledContent:
          type: string
        tone:
          type: string
        language:
          type: string
        emphasis:
          type: array
          items:
            type: string
    PopulatedScript:
      type: object
      properties:
        id:
          type: string
        sections:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [intro, weather, news, traffic, outro]
              content:
                type: string
              duration:
                type: integer
    TeleprompterScript:
      type: object
      properties:
        id:
          type: string
        presenterId:
          type: string
        createdAt:
          type: string
          format: date-time
        chunks:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              emphasis:
                type: boolean
              pause:
                type: integer
              notes:
                type: string
...